Update OperationInternal from Azure.Core
