+++ src/assets/Azure.Core.Shared/OperationInternal.cs
@@ -21,24 +21,29 @@ namespace Azure.Core
     /// </list>
     /// Supported members:
     /// <list type="bullet">
-    ///   <item><see cref="HasCompleted"/></item>
-    ///   <item><see cref="RawResponse"/>, used for <see cref="Operation.GetRawResponse"/></item>
-    ///   <item><see cref="UpdateStatus"/></item>
-    ///   <item><see cref="UpdateStatusAsync(CancellationToken)"/></item>
-    ///   <item><see cref="WaitForCompletionResponseAsync(CancellationToken)"/></item>
-    ///   <item><see cref="WaitForCompletionResponseAsync(TimeSpan, CancellationToken)"/></item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.HasCompleted"/></description>
+    ///   </item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.RawResponse"/>, used for <see cref="Operation.GetRawResponse"/></description>
+    ///   </item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.UpdateStatus"/></description>
+    ///   </item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.UpdateStatusAsync(CancellationToken)"/></description>
+    ///   </item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.WaitForCompletionResponseAsync(CancellationToken)"/></description>
+    ///   </item>
+    ///   <item>
+    ///     <description><see cref="OperationInternalBase.WaitForCompletionResponseAsync(TimeSpan, CancellationToken)"/></description>
+    ///   </item>
     /// </list>
     /// </summary>
-    internal class OperationInternal
+    internal class OperationInternal : OperationInternalBase
     {
-        private readonly IOperation? _operation;
-        private readonly ClientDiagnostics _diagnostics;
-        private readonly string _updateStatusScopeName;
-        private readonly IReadOnlyDictionary<string, string>? _scopeAttributes;
-        private RequestFailedException _operationFailedException = null!;
-        private const string RetryAfterHeaderName = "Retry-After";
-        private const string RetryAfterMsHeaderName = "retry-after-ms";
-        private const string XRetryAfterMsHeaderName = "x-ms-retry-after-ms";
+        private readonly IOperation _operation;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="OperationInternal"/> class.
@@ -46,7 +51,7 @@ namespace Azure.Core
         /// <param name="clientDiagnostics">Used for diagnostic scope and exception creation. This is expected to be the instance created during the construction of your main client.</param>
         /// <param name="operation">The long-running operation making use of this class. Passing "<c>this</c>" is expected.</param>
         /// <param name="rawResponse">
-        /// The initial value of <see cref="RawResponse"/>. Usually, long-running operation objects can be instantiated in two ways:
+        /// The initial value of <see cref="OperationInternalBase.RawResponse"/>. Usually, long-running operation objects can be instantiated in two ways:
         /// <list type="bullet">
         ///   <item>
         ///   When calling a client's "<c>Start&lt;OperationName&gt;</c>" method, a service call is made to start the operation, and an <see cref="Operation"/> instance is returned.
@@ -64,219 +69,19 @@ namespace Azure.Core
         /// <param name="scopeAttributes">The attributes to use during diagnostic scope creation.</param>
         public OperationInternal(
             ClientDiagnostics clientDiagnostics,
-            IOperation? operation,
+            IOperation operation,
             Response rawResponse,
             string? operationTypeName = null,
             IEnumerable<KeyValuePair<string, string>>? scopeAttributes = null)
+            :base(clientDiagnostics, rawResponse, operationTypeName ?? operation.GetType().Name, scopeAttributes)
         {
-            operationTypeName ??= operation?.GetType().Name;
-
             _operation = operation;
-            _diagnostics = clientDiagnostics;
-            _updateStatusScopeName = $"{operationTypeName}.UpdateStatus";
-            _scopeAttributes = scopeAttributes?.ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
-            RawResponse = rawResponse;
-            DefaultPollingInterval = TimeSpan.FromSeconds(1);
         }
 
-        /// <summary>
-        /// Returns <c>true</c> if the long-running operation has completed.
-        /// <example>Usage example:
-        /// <code>
-        ///   public bool HasCompleted => _operationInternal.HasCompleted;
-        /// </code>
-        /// </example>
-        /// </summary>
-        public bool HasCompleted { get; protected set; }
-
-        /// <summary>
-        /// The last HTTP response received from the server. Its update already handled in calls to "<c>UpdateStatus</c>" and
-        /// "<c>WaitForCompletionAsync</c>", but custom methods not supported by this class, such as "<c>CancelOperation</c>",
-        /// must update it as well.
-        /// <example>Usage example:
-        /// <code>
-        ///   public Response GetRawResponse() => _operationInternal.RawResponse;
-        /// </code>
-        /// </example>
-        /// </summary>
-        public Response RawResponse { get; set; }
-
-        /// <summary>
-        /// Can be set to control the default interval used between service calls in <see cref="WaitForCompletionResponseAsync(CancellationToken)"/>.
-        /// Defaults to 1 second.
-        /// </summary>
-        public TimeSpan DefaultPollingInterval { get; set; }
-
-        /// <summary>
-        /// Calls the server to get the latest status of the long-running operation, handling diagnostic scope creation for distributed
-        /// tracing. The default scope name can be changed with the "<c>operationTypeName</c>" parameter passed to the constructor.
-        /// <example>Usage example:
-        /// <code>
-        ///   public async ValueTask&lt;Response&gt; UpdateStatusAsync(CancellationToken cancellationToken) =>
-        ///     await _operationInternal.UpdateStatusAsync(cancellationToken).ConfigureAwait(false);
-        /// </code>
-        /// </example>
-        /// </summary>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
-        /// <returns>The HTTP response received from the server.</returns>
-        /// <remarks>
-        /// After a successful run, this method will update <see cref="RawResponse"/> and might update <see cref="HasCompleted"/>.
-        /// </remarks>
-        /// <exception cref="RequestFailedException">Thrown if there's been any issues during the connection, or if the operation has completed with failures.</exception>
-        public async ValueTask<Response> UpdateStatusAsync(CancellationToken cancellationToken) =>
-            await UpdateStatusAsync(async: true, cancellationToken).ConfigureAwait(false);
-
-        /// <summary>
-        /// Calls the server to get the latest status of the long-running operation, handling diagnostic scope creation for distributed
-        /// tracing. The default scope name can be changed with the "<c>operationTypeName</c>" parameter passed to the constructor.
-        /// <example>Usage example:
-        /// <code>
-        ///   public Response UpdateStatus(CancellationToken cancellationToken) => _operationInternal.UpdateStatus(cancellationToken);
-        /// </code>
-        /// </example>
-        /// </summary>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
-        /// <returns>The HTTP response received from the server.</returns>
-        /// <remarks>
-        /// After a successful run, this method will update <see cref="RawResponse"/> and might update <see cref="HasCompleted"/>.
-        /// </remarks>
-        /// <exception cref="RequestFailedException">Thrown if there's been any issues during the connection, or if the operation has completed with failures.</exception>
-        public Response UpdateStatus(CancellationToken cancellationToken) =>
-            UpdateStatusAsync(async: false, cancellationToken).EnsureCompleted();
-
-        /// <summary>
-        /// Periodically calls <see cref="UpdateStatusAsync(CancellationToken)"/> until the long-running operation completes. The interval
-        /// between calls is defined by the property <see cref="DefaultPollingInterval"/>, but it can change based on information returned
-        /// from the server. After each service call, a retry-after header may be returned to communicate that there is no reason to poll
-        /// for status change until the specified time has passed. In this case, the maximum value between the <see cref="DefaultPollingInterval"/>
-        /// property and the retry-after header is chosen as the wait interval. Headers supported are: "Retry-After", "retry-after-ms",
-        /// and "x-ms-retry-after-ms".
-        /// <example>Usage example:
-        /// <code>
-        ///   public async ValueTask&lt;Response&lt;T&gt;&gt; WaitForCompletionAsync(CancellationToken cancellationToken) =>
-        ///     await _operationInternal.WaitForCompletionAsync(cancellationToken).ConfigureAwait(false);
-        /// </code>
-        /// </example>
-        /// </summary>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
-        /// <returns>The last HTTP response received from the server, including the final result of the long-running operation.</returns>
-        /// <exception cref="RequestFailedException">Thrown if there's been any issues during the connection, or if the operation has completed with failures.</exception>
-        public virtual async ValueTask<Response> WaitForCompletionResponseAsync(CancellationToken cancellationToken) =>
-            await WaitForCompletionResponseAsync(DefaultPollingInterval, cancellationToken).ConfigureAwait(false);
-
-        /// <summary>
-        /// Periodically calls <see cref="UpdateStatusAsync(CancellationToken)"/> until the long-running operation completes. The interval
-        /// between calls is defined by the parameter <paramref name="pollingInterval"/>, but it can change based on information returned
-        /// from the server. After each service call, a retry-after header may be returned to communicate that there is no reason to poll
-        /// for status change until the specified time has passed. In this case, the maximum value between the <paramref name="pollingInterval"/>
-        /// parameter and the retry-after header is chosen as the wait interval. Headers supported are: "Retry-After", "retry-after-ms",
-        /// and "x-ms-retry-after-ms".
-        /// <example>Usage example:
-        /// <code>
-        ///   public async ValueTask&lt;Response&lt;T&gt;&gt; WaitForCompletionAsync(TimeSpan pollingInterval, CancellationToken cancellationToken) =>
-        ///     await _operationInternal.WaitForCompletionAsync(pollingInterval, cancellationToken).ConfigureAwait(false);
-        /// </code>
-        /// </example>
-        /// </summary>
-        /// <param name="pollingInterval">The interval between status requests to the server.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
-        /// <returns>The last HTTP response received from the server, including the final result of the long-running operation.</returns>
-        /// <exception cref="RequestFailedException">Thrown if there's been any issues during the connection, or if the operation has completed with failures.</exception>
-        public virtual async ValueTask<Response> WaitForCompletionResponseAsync(TimeSpan pollingInterval, CancellationToken cancellationToken)
+        protected override async ValueTask<Response> UpdateStateAsync(bool async, CancellationToken cancellationToken)
         {
-            while (true)
-            {
-                Response response = await UpdateStatusAsync(cancellationToken).ConfigureAwait(false);
-
-                if (HasCompleted)
-                {
-                    return response;
-                }
-
-                TimeSpan delay = GetServerDelay(response, pollingInterval);
-                await WaitAsync(delay, cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        protected virtual async Task WaitAsync(TimeSpan delay, CancellationToken cancellationToken) =>
-            await Task.Delay(delay, cancellationToken).ConfigureAwait(false);
-
-        private async ValueTask<Response> UpdateStatusAsync(bool async, CancellationToken cancellationToken)
-        {
-            using DiagnosticScope scope = _diagnostics.CreateScope(_updateStatusScopeName);
-
-            if (_scopeAttributes != null)
-            {
-                foreach (KeyValuePair<string, string> attribute in _scopeAttributes)
-                {
-                    scope.AddAttribute(attribute.Key, attribute.Value);
-                }
-            }
-
-            scope.Start();
-
-            try
-            {
-                return await UpdateStateAsync(async, cancellationToken).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                scope.Failed(e);
-                throw;
-            }
-        }
-
-        protected virtual async ValueTask<Response> UpdateStateAsync(bool async, CancellationToken cancellationToken)
-        {
-            OperationState state = await _operation!.UpdateStateAsync(async, cancellationToken).ConfigureAwait(false);
-
-            RawResponse = state.RawResponse;
-
-            if (state.HasCompleted)
-            {
-                if (state.HasSucceeded)
-                {
-                    HasCompleted = true;
-                }
-                else
-                {
-                    _operationFailedException = state.OperationFailedException ??
-                                                (async
-                                                    ? await _diagnostics
-                                                        .CreateRequestFailedExceptionAsync(state.RawResponse)
-                                                        .ConfigureAwait(false)
-                                                    : _diagnostics.CreateRequestFailedException(state.RawResponse));
-                    HasCompleted = true;
-
-                    throw _operationFailedException;
-                }
-            }
-
-            return state.RawResponse;
-        }
-
-        protected static TimeSpan GetServerDelay(Response response, TimeSpan pollingInterval)
-        {
-            TimeSpan serverDelay = pollingInterval;
-            if (response.Headers.TryGetValue(RetryAfterMsHeaderName, out string? retryAfterValue) ||
-                response.Headers.TryGetValue(XRetryAfterMsHeaderName, out retryAfterValue))
-            {
-                if (int.TryParse(retryAfterValue, out int serverDelayInMilliseconds))
-                {
-                    serverDelay = TimeSpan.FromMilliseconds(serverDelayInMilliseconds);
-                }
-            }
-            else if (response.Headers.TryGetValue(RetryAfterHeaderName, out retryAfterValue))
-            {
-                if (int.TryParse(retryAfterValue, out int serverDelayInSeconds))
-                {
-                    serverDelay = TimeSpan.FromSeconds(serverDelayInSeconds);
-                }
-            }
-
-            return serverDelay > pollingInterval
-                ? serverDelay
-                : pollingInterval;
+            OperationState state = await _operation.UpdateStateAsync(async, cancellationToken).ConfigureAwait(false);
+            return await ApplyStateAsync(async, state.RawResponse, state.HasCompleted, state.HasSucceeded, state.OperationFailedException).ConfigureAwait(false);
         }
     }
 
@@ -288,7 +93,7 @@ namespace Azure.Core
     {
         /// <summary>
         /// Calls the service and updates the state of the long-running operation. Properties directly handled by the
-        /// <see cref="OperationInternal"/> class, such as <see cref="OperationInternal.RawResponse"/>
+        /// <see cref="OperationInternal"/> class, such as <see cref="OperationInternalBase.RawResponse"/>
         /// don't need to be updated. Operation-specific properties, such as "<c>CreateOn</c>" or "<c>LastModified</c>",
         /// must be manually updated by the operation implementing this method.
         /// <example>Usage example:
